add player.add_piece() and remove_piece()
Adds two functions to the player class to add and remove pieces from the
board.
Changed get_center() to get_tile(), and made it a method of the player
class.
Changed Player methods to return True or False depending on if the
operation was successful or not. Currently, success is simply predicated
on if there are enough pieces/ room in the reserve.