add Board class
Adds a Board class that represents the state of the game board. The main
event loop will capture key / mouse input, perform the necessary logic
of where the mouse cursor is etc., then command the Board to make a
particular change.

Clicking above and below the center line of the board will add/ remove
reserve pieces for the top and bottom player, respectively.